// src/types/enum.ts
export enum ROOT_PROPERTY_NAME {
  BACKGROUND = "--background",
  FOREGROUND = "--foreground",

  CARD = "--card",
  CARD_FOREGROUND = "--card-foreground",

  POPOVER = "--popover",
  POPOVER_FOREGROUND = "--popover-foreground",

  PRIMARY = "--primary",
  PRIMARY_FOREGROUND = "--primary-foreground",

  SECONDARY = "--secondary",
  SECONDARY_FOREGROUND = "--secondary-foreground",

  MUTED = "--muted",
  MUTED_FOREGROUND = "--muted-foreground",

  ACCENT = "--accent",
  ACCENT_FOREGROUND = "--accent-foreground",

  DESTRUCTIVE = "--destructive",
  DESTRUCTIVE_FOREGROUND = "--destructive-foreground",

  BORDER = "--border",
  INPUT = "--input",
  RING = "--ring",

  RADIUS = "--radius",
}